+++ db/migrate/20180519204655_add_fixed_to_alchemy_elements.rb
@@ -1,6 +1,0 @@
-class AddFixedToAlchemyElements < ActiveRecord::Migration[5.0]
-  def change
-    add_column :alchemy_elements, :fixed, :boolean, default: false, null: false
-    add_index :alchemy_elements, :fixed
-  end
-end