+++ db/migrate/20191016073858_create_alchemy_essence_pages.rb
@@ -1,8 +1,0 @@
-class CreateAlchemyEssencePages < ActiveRecord::Migration[5.0]
-  def change
-    create_table :alchemy_essence_pages do |t|
-      t.references :page, null: true, foreign_key: { to_table: :alchemy_pages }
-      t.timestamps
-    end
-  end
-end