+++ spec/dummy/db/migrate/20180226123013_alchemy_four_point_zero.rb
@@ -1,1 +1,0 @@
-../../../../db/migrate/20180226123013_alchemy_four_point_zero.rb