+++ spec/dummy/db/migrate/20200226081535_add_site_id_to_alchemy_nodes.rb
@@ -1,1 +1,0 @@
-../../../../db/migrate/20200226081535_add_site_id_to_alchemy_nodes.rb